bisect: move even the option parsing to `bisect--helper`
On our journey to a fully built-in `git bisect`, this is the
second-to-last step.

Side note: The `if (!strcmp(...)) ... else if (!strcmp(...)) ... else if
(!strcmp(...)) ...` chain seen in this patch was not actually the first
idea how to convert the command modes to sub-commands. Since the
`bisect--helper` command already used the `parse-opions` API with neatly
set-up command modes, naturally the idea was to use `PARSE_OPT_NODASH`
to support proper sub-commands instead. However, the `parse-options` API
is not set up for that, and even after making that option work with long
options, it turned out that `STOP_AT_NON_OPTION` and `KEEP_UNKNOWN`
would have to be used but these options were not designed to work
together. So it would appear as if a lot of work would have to be done
just to be able to use `parse_options()` just to parse the sub-command,
instead of a simple `if...else if` chain, the latter being a
dramatically simpler implementation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>