repo-settings: name the default fetch.negotiationAlgorithm 'consecutive'
Give the default fetch.negotiationAlgorithm the name 'consecutive' and
update the documentation accordingly.  Since there may be some users
using the name 'default' for this behavior, retain that name for now.
We do not want to use that name indefinitely, though, because if
'skipping' becomes the default, then the "default" behavior will not be
the default behavior, which would be confusing.

Signed-off-by: Elijah Newren <newren@gmail.com>